migrate channel types customizations to class override mechanism.
